Add a delete_tag option
If this option is set then it specifies a notmuch tag that will be
applied on unlink instead of unlinking the actual message file.  This,
combined with the search.exclude_tags option allows for deleting mail
from the FUSE filesystem without actually losing any mail.

With fixes from pngwjpgh & tsto.

Fixes #11 and fixes #8.